+++ lizard_progress/urls.py
@@ -8,6 +8,8 @@ from lizard_ui.urls import debugmode_urlpatterns
 from lizard_progress.views import View
 from lizard_progress.views import MapView
 from lizard_progress.views import UploadView
+from lizard_progress.views import ComparisonView
+from lizard_progress.views import ComparisonPopupView
 from lizard_progress.views import DashboardView
 from lizard_progress.views import DashboardCsvView
 from lizard_progress.views import DashboardAreaView
@@ -22,6 +24,15 @@ urlpatterns = patterns(
         name='lizard_progress_view'),
     url('^(?P<project_slug>[^/]+)/map/$', login_required(MapView.as_view()),
         name='lizard_progress_mapview'),
+    url('^(?P<project_slug>[^/]+)/comparison/$', login_required(ComparisonView.as_view()),
+        name='lizard_progress_comparisonview'),
+    url('^(?P<project_slug>[^/]+)/comparison/(?P<mtype_slug>[^/]+)/$',
+        login_required(ComparisonView.as_view()),
+        name='lizard_progress_comparisonview2'),
+    url('^(?P<project_slug>[^/]+)/comparison/popup/'
+        '(?P<mtype_slug>[^/]+)/(?P<location_unique_id>[^/]+)/$',
+        login_required(ComparisonPopupView.as_view()),
+        name='lizard_progress_comparisonpopup'),
     url('^(?P<project_slug>[^/]+)/dashboard/$',
         login_required(DashboardView.as_view()),
         name='lizard_progress_dashboardview'),