+++ lizard_progress/views.py
@@ -8,6 +8,8 @@ UploadView - where users can upload files
 DashboardAreaView - a graph of the project's progress (hence "lizard-progress")
 DashboardCsvView - a csv file view
 protected_file_download - to download some uploaded files
+ComparisonView - a listing of measurements made by more than one
+                 contractor, that can be shown side by side
 """
 
 import csv
@@ -34,12 +36,14 @@ from django.views.static import serve
 from lizard_map.matplotlib_settings import SCREEN_DPI
 from lizard_map.views import AppView
 from lizard_progress import specifics
-from lizard_progress.models import has_access
+from lizard_progress.layers import ProgressAdapter
 from lizard_progress.models import Area
 from lizard_progress.models import Contractor
-from lizard_progress.models import Project
 from lizard_progress.models import MeasurementType
+from lizard_progress.models import Location
+from lizard_progress.models import Project
 from lizard_progress.models import ScheduledMeasurement
+from lizard_progress.models import has_access
 from lizard_progress.tools import unique_filename
 from lizard_ui.views import ViewContextMixin
 
@@ -110,6 +114,10 @@ class View(AppView):
         self.project = get_object_or_404(Project, slug=self.project_slug)
 
         if has_access(request.user, self.project):
+            self.has_full_access = all(
+                has_access(request.user, self.project, contractor)
+                for contractor in self.project.contractor_set.all())
+
             return super(View, self).dispatch(request, *args, **kwargs)
         else:
             raise PermissionDenied()
@@ -127,6 +135,11 @@ class View(AppView):
     def map_url(self):
         """Returns URL to this project's Map view"""
         return reverse('lizard_progress_mapview',
+                       kwargs={'project_slug': self.project_slug})
+
+    def comparison_url(self):
+        """Returns URL to this project's Comparison view"""
+        return reverse('lizard_progress_comparisonview',
                        kwargs={'project_slug': self.project_slug})
 
 
@@ -180,6 +193,230 @@ class MapView(View):
                                         self.project.slug}),
                             })
         return layers
+
+
+class ComparisonView(View):
+    """View that can show measurement types of this project, and if one is chosen,
+    a sorted list of locations with more than one contractor."""
+
+    template_name = 'lizard_progress/comparison.html'
+
+    def dispatch(self, request, *args, **kwargs):
+        """Check access (user needs to be able to read data of all contractors),
+        and find the current measurement type, if any."""
+
+        self.mtype_slug = kwargs.get('mtype_slug', None)
+
+        result = super(ComparisonView, self).dispatch(request, *args, **kwargs)
+
+        for contractor in self.project.contractor_set.all():
+            if not has_access(request.user, self.project, contractor):
+                raise PermisionDenied()
+
+        return result
+
+    def crumbs(self):
+        """Breadcrumb for this page."""
+        crumbs = super(ComparisonView, self).crumbs()
+
+        crumbs.append({
+                'url': self.comparison_url(),
+                'description': 'Vergelijking',
+                'title': '%s vergelijking' % (self.project.name,)
+                })
+
+        return crumbs
+
+    @property
+    def measurement_type(self):
+        if self.mtype_slug is not None:
+            return MeasurementType.objects.get(slug=self.mtype_slug)
+        else:
+            return None
+
+    def measurement_types(self):
+        """Return available measurement types."""
+
+        measurement_types = self.project.measurementtype_set.all()
+        logger.debug("Hrm.")
+        logger.debug(measurement_types)
+        return measurement_types
+
+    def measurement_types_urls(self):
+        return [(measurement_type,
+                 reverse(
+                    'lizard_progress_comparisonview2',
+                    kwargs={
+                        'project_slug': self.project.slug,
+                        'mtype_slug': measurement_type.slug,
+                        }))
+                for measurement_type in self.measurement_types()]
+
+    def locations_to_compare(self):
+        """Locations that have more than scheduled measurements by
+        more than one contractor, for this measurement type"""
+
+        measurement_type = self.measurement_type
+
+        if not measurement_type or not measurement_type.id:
+            return ()
+
+        # We're looking for locations in this project where the number
+        # of distinct contractors that have a completed scheduled
+        # measurement of this measurement type on that location is
+        # greater than 1.
+
+        # As far as I can see, that beats Django's ORM and we need to
+        # use SQL.
+        locations = (
+            Location.objects.filter(project=self.project).
+            extra(where=["""
+         (SELECT
+            COUNT(
+              DISTINCT lizard_progress_scheduledmeasurement.contractor_id)
+          FROM
+            lizard_progress_scheduledmeasurement
+          WHERE
+            lizard_progress_scheduledmeasurement.location_id =
+              lizard_progress_location.unique_id
+          AND
+            lizard_progress_scheduledmeasurement.complete='t'
+          AND
+            lizard_progress_scheduledmeasurement.measurement_type_id=%d
+          ) > 1
+         """ % measurement_type.id]).
+            order_by('unique_id').
+            all())
+
+        return [(location, self.comparison_popup_url(measurement_type, location))
+                for location in locations]
+
+    def comparison_popup_url(self, measurement_type, location):
+        """Returns URL to this project's Comparison view"""
+        return reverse(
+            'lizard_progress_comparisonpopup',
+            kwargs={
+                'project_slug': self.project_slug,
+                'mtype_slug': measurement_type.slug,
+                'location_unique_id': location.unique_id,
+                }
+            )
+
+
+class ComparisonPopupView(View):
+    template_name = 'lizard_progress/comparison_popup.html'
+
+    def dispatch(self, request, *args, **kwargs):
+        """Check access (user needs to be able to read data of all contractors),
+        and find the current measurement type, if any."""
+
+        self.mtype_slug = kwargs.get('mtype_slug', None)
+        self.measurement_type = MeasurementType.objects.get(slug=self.mtype_slug)
+
+        self.location_unique_id = kwargs.get('location_unique_id', None)
+        self.location = Location.objects.get(unique_id=self.location_unique_id)
+
+        result = super(ComparisonPopupView, self).dispatch(request, *args, **kwargs)
+
+        for contractor in self.project.contractor_set.all():
+            if not has_access(request.user, self.project, contractor):
+                raise PermisionDenied()
+
+        return result
+
+    def contractors(self):
+        """Return the contractors that have a complete measurement at this location"""
+
+        contractors = set()
+
+        scheduled_measurements = (ScheduledMeasurement.objects.
+                                  filter(project=self.project).
+                                  filter(measurement_type=self.measurement_type).
+                                  filter(location=self.location).
+                                  filter(complete=True))
+
+        for sm in scheduled_measurements:
+            contractors.add(sm.contractor)
+
+        return sorted(contractors,
+                      cmp=lambda a,b: cmp(a.name, b.name))
+
+
+    def contractor_html(self):
+        """Perform lizard-map-ish magic to get the same HTML as normal popups do."""
+        htmls = []
+
+        class FakeWorkspaceItem(object):
+            adapter_class = 'adapter_progress'
+
+            def __init__(self, layer_arguments):
+                self.adapter_layer_json=json.dumps(layer_arguments)
+
+            def _url_arguments(self, identifiers):
+                """for img_url, csv_url"""
+
+                from lizard_map.adapter import adapter_serialize
+
+                layer_json = self.adapter_layer_json.replace('"', '%22')
+                url_arguments = [
+                    'adapter_layer_json=%s' % layer_json, ]
+                url_arguments.extend([
+                        'identifier=%s' % adapter_serialize(
+                            identifier) for identifier in identifiers])
+                return url_arguments
+
+            def url(self, url_name, identifiers, extra_kwargs=None):
+                """fetch url to adapter (img, csv, ...)
+
+                example url_name: "lizard_map_adapter_image"
+                """
+                kwargs = {'adapter_class': 'adapter_progress'}
+                if extra_kwargs is not None:
+                    kwargs.update(extra_kwargs)
+                url = reverse(
+                    url_name,
+                    kwargs=kwargs,
+                    )
+                url += '?' + '&'.join(self._url_arguments(identifiers))
+                return url
+
+        for contractor in self.contractors():
+            try:
+                layer_arguments={
+                    'project_slug': self.project.slug,
+                    'contractor_slug': contractor.slug,
+                    'measurement_type_slug': self.measurement_type.slug,
+                    }
+
+                workspace_item = FakeWorkspaceItem(layer_arguments)
+
+                adapter = ProgressAdapter(
+                    workspace_item=workspace_item,
+                    layer_arguments=layer_arguments,
+                    )
+
+            except Exception as e:
+                logger.critical("Adapter exception: "+str(e))
+
+            scheduled_measurements = (
+                ScheduledMeasurement.objects.
+                filter(project=self.project).
+                filter(location=self.location).
+                filter(contractor=contractor).
+                filter(measurement_type=self.measurement_type))
+
+            sm_ids = [{
+                    'scheduled_measurement_id': sm.id,
+                    } for sm in scheduled_measurements]
+
+            logger.critical("SM IDs: "+str(sm_ids))
+
+            try:
+                htmls.append(adapter.html(identifiers=sm_ids))
+            except Exception as e:
+                logger.critical("HTML exception: "+str(e))
+
+        return htmls
 
 
 class DashboardView(View):